requirements.txt
alembic==1.6.5
atomicwrites==1.4.0
attrs==20.3.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.5
dnspython==2.1.0
email-validator==1.1.3
Flask==1.1.2
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.18
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
